fix(reabastecimiento): cantidades enteras redondeadas hacia arriba
La validacion aceptaba decimales: z.number().positive() dejaba pasar 2,5 y
ese decimal llegaba hasta la linea de la orden de compra en Odoo. Nadie le
pide 2,5 unidades a un proveedor.

Se corrige en el borde (zod exige entero) y en el input (Math.ceil sobre lo
que se escribe, no solo al mostrar), no solo en la presentacion.

Stock y dias de stock pasan de Math.round/toFixed a Math.ceil.

Excepcion deliberada: 'vende/dia' conserva un decimal. Es una tasa, no una
cantidad — redondear 1,4 a 2 diria que un producto vende 43% mas de lo que
vende, y esa cifra alimenta la lectura de la velocidad.

+++ src/app/(dashboard)/reabastecimiento/actions.ts
@@ -25,8 +25,8 @@ const approveSchema = z.object({
       z.object({
         odooProductId: z.number().int().positive(),
         productName: z.string().min(1).max(300),
-        qty: z.number().positive().max(100_000),
-        suggestedQty: z.number().nonnegative(),
+        qty: z.number().int().positive().max(100_000), // unidades enteras: no se piden 2,5
+        suggestedQty: z.number().int().nonnegative(),
         unitCost: z.number().nonnegative(),
         reason: z.enum(["critico", "advertencia", "min_stock"]),
       })
